Fix nested any/all evaluation
The current evaluator pops everything on the stack in order to evaluate `any` or `all` queries. That breaks in nested situations like:

```
all(any(unix, target_arch="x86"), not(any(target_os="android", target_os="emscripten")))
```

Fix this by tracking the number of predicates in the `Any` and `All` variants.

This unfortunately breaks semver because the definition of `Func` changes.
`Func` doesn't seem to be exported anywhere so maybe in the upcoming release
it could be made private?