Implement evaluation when the result is unknown
I've run into some situations where the result of evaluating a predicate is
unknown, most notably with target families. Allow evaluating such cases
using a simple three-valued logic.

Do this by making the logic a trait -- its implementation for `bool` is the
standard boolean logic, and its implementation for `Option<bool>` is the
Kleene K3 logic. This logic is also used in SQL.