Build with Rust 1.52
This crate is a dependency of `system-deps` which is a pretty
foundational library.  In my case, we use the `glib` and gtk-rs bindings
which transitively depend on this.

Not everyone can move at upstream Rust speed.  While there are improvements,
there are sometimes regressions too (thankfully, not many, but still).

In this case, I hope you'd agree that while the new "or pattern" feature
is *nice* it's far from a must have.  (Unlike e.g. const generics or so)