[RFC] support panic and target_has_atomic predicates
There are a couple of new predicates in Rust 1.60: `target_has_atomic`
and `panic`. Here's a first cut at supporting them.

The panic predicate depends both on native platform support (default
behavior) and on a potential override through RUSTFLAGS. This can be
handled in downstream libraries that need such overrides.